requirements.txt
aiosqlite
fastapi
jinja2
pyrogram
tgcrypto
uvicorn[standard]
python-dotenv
aiofiles
